back/requirements.txt
asgiref==3.5.2
dj-rest-auth==2.2.5
django-environ==0.9.0
django-filter==22.1
Django==4.1.2
djangorestframework-camel-case==1.3.0
djangorestframework==3.14.0
pip==22.1.2
psycopg2==2.9.4
pytz==2022.4
setuptools==62.6.0
sqlparse==0.4.3
wheel==0.37.1
